+++ protocols/kad/src/behaviour.rs
@@ -179,23 +179,6 @@ pub struct KademliaConfig {
     caching: KademliaCaching,
 }
 
-/// The configuration for Kademlia "write-back" caching after successful
-/// lookups via [`Kademlia::get_record`].
-#[derive(Debug, Clone)]
-pub enum KademliaCaching {
-    /// Caching is disabled and the peers closest to records being looked up
-    /// that do not return a record are not tracked, i.e.
-    /// [`GetRecordOk::cache_candidates`] is always empty.
-    Disabled,
-    /// Up to `max_peers` peers not returning a record that are closest to the key
-    /// being looked up are tracked and returned in [`GetRecordOk::cache_candidates`].
-    /// Furthermore, if [`Kademlia::get_record`] is used with a quorum of 1, the
-    /// found record is automatically sent to (i.e. cached at) these peers. For lookups with a
-    /// quorum > 1, the write-back operation must be performed explicitly, if
-    /// desired and after choosing a record from the results, via [`Kademlia::put_record_to`].
-    Enabled { max_peers: u16 },
-}
-
 impl Default for KademliaConfig {
     fn default() -> Self {
         KademliaConfig {
@@ -213,6 +196,21 @@ impl Default for KademliaConfig {
             caching: KademliaCaching::Enabled { max_peers: 1 },
         }
     }
+}
+
+/// The configuration for Kademlia "write-back" caching after successful
+/// lookups via [`Kademlia::get_record`].
+#[derive(Debug, Clone)]
+pub enum KademliaCaching {
+    /// Caching is disabled and the peers closest to records being looked up
+    /// that do not return a record are not tracked, i.e.
+    /// [`GetRecordOk::FinishedWithNoAdditionalRecord`] is always empty.
+    Disabled,
+    /// Up to `max_peers` peers not returning a record that are closest to the key
+    /// being looked up are tracked and returned in [`GetRecordOk::FinishedWithNoAdditionalRecord`].
+    /// The write-back operation must be performed explicitly, if
+    /// desired and after choosing a record from the results, via [`Kademlia::put_record_to`].
+    Enabled { max_peers: u16 },
 }
 
 impl KademliaConfig {
@@ -435,6 +433,7 @@ where
 
         Kademlia {
             store,
+            caching: config.caching,
             kbuckets: KBucketsTable::new(local_key, config.kbucket_pending_timeout),
             kbucket_inserts: config.kbucket_inserts,
             protocol_config: config.protocol_config,
@@ -448,7 +447,6 @@ where
             provider_record_ttl: config.provider_record_ttl,
             connection_idle_timeout: config.connection_idle_timeout,
             local_addrs: HashSet::new(),
-            caching: config.caching,
         }
     }
 
@@ -661,13 +659,15 @@ where
     where
         K: Into<kbucket::Key<K>> + Into<Vec<u8>> + Clone,
     {
+        let target: kbucket::Key<K> = key.clone().into();
+        let key: Vec<u8> = key.into();
         let info = QueryInfo::GetClosestPeers {
-            key: key.clone().into(),
+            key,
+            step: ProgressStep::first(),
         };
-        let target: kbucket::Key<K> = key.into();
-        let peers = self.kbuckets.closest_keys(&target);
+        let peer_keys: Vec<kbucket::Key<PeerId>> = self.kbuckets.closest_keys(&target).collect();
         let inner = QueryInner::new(info);
-        self.queries.add_iter_closest(target.clone(), peers, inner)
+        self.queries.add_iter_closest(target, peer_keys, inner)
     }
 
     /// Returns closest peers to the given key; takes peers from local routing table only.
@@ -682,36 +682,56 @@ where
     ///
     /// The result of this operation is delivered in a
     /// [`KademliaEvent::OutboundQueryCompleted{QueryResult::GetRecord}`].
-    pub fn get_record(&mut self, key: record::Key, quorum: Quorum) -> QueryId {
-        let quorum = quorum.eval(self.queries.config().replication_factor);
-        let mut records = Vec::with_capacity(quorum.get());
-
-        if let Some(record) = self.store.get(&key) {
+    pub fn get_record(&mut self, key: record::Key) -> QueryId {
+        let record = if let Some(record) = self.store.get(&key) {
             if record.is_expired(Instant::now()) {
-                self.store.remove(&key)
+                self.store.remove(&key);
+                None
             } else {
-                records.push(PeerRecord {
+                Some(PeerRecord {
                     peer: None,
                     record: record.into_owned(),
-                });
+                })
             }
-        }
+        } else {
+            None
+        };
 
-        let done = records.len() >= quorum.get();
+        let step = ProgressStep::first();
+
         let target = kbucket::Key::new(key.clone());
-        let info = QueryInfo::GetRecord {
-            key,
-            records,
-            quorum,
-            cache_candidates: BTreeMap::new(),
+        let info = if record.is_some() {
+            QueryInfo::GetRecord {
+                key,
+                step: step.next(),
+                found_a_record: true,
+                cache_candidates: BTreeMap::new(),
+            }
+        } else {
+            QueryInfo::GetRecord {
+                key,
+                step: step.clone(),
+                found_a_record: false,
+                cache_candidates: BTreeMap::new(),
+            }
         };
         let peers = self.kbuckets.closest_keys(&target);
         let inner = QueryInner::new(info);
-        let id = self.queries.add_iter_closest(target.clone(), peers, inner); // (*)
+        let id = self.queries.add_iter_closest(target.clone(), peers, inner);
 
-        // Instantly finish the query if we already have enough records.
-        if done {
-            self.queries.get_mut(&id).expect("by (*)").finish();
+        // No queries were actually done for the results yet.
+        let stats = QueryStats::empty();
+
+        if let Some(record) = record {
+            self.queued_events
+                .push_back(NetworkBehaviourAction::GenerateEvent(
+                    KademliaEvent::OutboundQueryProgressed {
+                        id,
+                        result: QueryResult::GetRecord(Ok(GetRecordOk::FoundRecord(record))),
+                        step,
+                        stats,
+                    },
+                ));
         }
 
         id
@@ -766,19 +786,16 @@ where
     ///
     /// If the record's expiration is `None`, the configured record TTL is used.
     ///
-    /// > **Note**: This is not a regular Kademlia DHT operation. It may be
+    /// > **Note**: This is not a regular Kademlia DHT operation. It needs to be
     /// > used to selectively update or store a record to specific peers
     /// > for the purpose of e.g. making sure these peers have the latest
     /// > "version" of a record or to "cache" a record at further peers
     /// > to increase the lookup success rate on the DHT for other peers.
     /// >
-    /// > In particular, if lookups are performed with a quorum > 1 multiple
-    /// > possibly different records may be returned and the standard Kademlia
+    /// > In particular, there is no automatic storing of records performed, and this
+    /// > method must be used to ensure the standard Kademlia
     /// > procedure of "caching" (i.e. storing) a found record at the closest
-    /// > node to the key that _did not_ return it cannot be employed
-    /// > transparently. In that case, client code can explicitly choose
-    /// > which record to store at which peers for analogous write-back
-    /// > caching or for other reasons.
+    /// > node to the key that _did not_ return it.
     pub fn put_record_to<I>(&mut self, mut record: Record, peers: I, quorum: Quorum) -> QueryId
     where
         I: ExactSizeIterator<Item = PeerId>,
@@ -854,6 +871,7 @@ where
         let info = QueryInfo::Bootstrap {
             peer: *local_key.preimage(),
             remaining: None,
+            step: ProgressStep::first(),
         };
         let peers = self.kbuckets.closest_keys(&local_key).collect::<Vec<_>>();
         if peers.is_empty() {
@@ -924,21 +942,49 @@ where
     /// The result of this operation is delivered in a
     /// reported via [`KademliaEvent::OutboundQueryCompleted{QueryResult::GetProviders}`].
     pub fn get_providers(&mut self, key: record::Key) -> QueryId {
-        let providers = self
+        let providers: HashSet<_> = self
             .store
             .providers(&key)
             .into_iter()
             .filter(|p| !p.is_expired(Instant::now()))
             .map(|p| p.provider)
             .collect();
+
+        let step = ProgressStep::first();
+
         let info = QueryInfo::GetProviders {
             key: key.clone(),
-            providers,
+            providers_found: providers.len(),
+            step: if providers.is_empty() {
+                step.clone()
+            } else {
+                step.next()
+            },
         };
-        let target = kbucket::Key::new(key);
+
+        let target = kbucket::Key::new(key.clone());
         let peers = self.kbuckets.closest_keys(&target);
         let inner = QueryInner::new(info);
-        self.queries.add_iter_closest(target.clone(), peers, inner)
+        let id = self.queries.add_iter_closest(target.clone(), peers, inner);
+
+        // No queries were actually done for the results yet.
+        let stats = QueryStats::empty();
+
+        if !providers.is_empty() {
+            self.queued_events
+                .push_back(NetworkBehaviourAction::GenerateEvent(
+                    KademliaEvent::OutboundQueryProgressed {
+                        id,
+                        result: QueryResult::GetProviders(Ok(GetProvidersOk::FoundProviders {
+                            key,
+                            providers,
+                        })),
+                        step,
+                        stats,
+                    },
+                ));
+        }
+        id
     }
 
     /// Processes discovered peers from a successful request in an iterative `Query`.
@@ -1189,7 +1235,11 @@ where
         log::trace!("Query {:?} finished.", query_id);
         let result = q.into_result();
         match result.inner.info {
-            QueryInfo::Bootstrap { peer, remaining } => {
+            QueryInfo::Bootstrap {
+                peer,
+                remaining,
+                mut step,
+            } => {
                 let local_key = self.kbuckets.local_key().clone();
                 let mut remaining = remaining.unwrap_or_else(|| {
                     debug_assert_eq!(&peer, local_key.preimage());
@@ -1235,41 +1285,53 @@ where
                     let info = QueryInfo::Bootstrap {
                         peer: *target.preimage(),
                         remaining: Some(remaining),
+                        step: step.next(),
                     };
                     let peers = self.kbuckets.closest_keys(&target);
                     let inner = QueryInner::new(info);
                     self.queries
                         .continue_iter_closest(query_id, target.clone(), peers, inner);
-                }
+                } else {
+                    step.last = true;
+                };
 
-                Some(KademliaEvent::OutboundQueryCompleted {
+                Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::Bootstrap(Ok(BootstrapOk {
                         peer,
                         num_remaining,
                     })),
+                    step,
                 })
             }
 
-            QueryInfo::GetClosestPeers { key, .. } => Some(KademliaEvent::OutboundQueryCompleted {
-                id: query_id,
-                stats: result.stats,
-                result: QueryResult::GetClosestPeers(Ok(GetClosestPeersOk {
-                    key,
-                    peers: result.peers.collect(),
-                })),
-            }),
+            QueryInfo::GetClosestPeers { key, mut step } => {
+                step.last = true;
 
-            QueryInfo::GetProviders { key, providers } => {
-                Some(KademliaEvent::OutboundQueryCompleted {
+                Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
-                    result: QueryResult::GetProviders(Ok(GetProvidersOk {
+                    result: QueryResult::GetClosestPeers(Ok(GetClosestPeersOk {
                         key,
-                        providers,
-                        closest_peers: result.peers.collect(),
+                        peers: result.peers.collect(),
                     })),
+                    step,
+                })
+            }
+
+            QueryInfo::GetProviders { mut step, .. } => {
+                step.last = true;
+
+                Some(KademliaEvent::OutboundQueryProgressed {
+                    id: query_id,
+                    stats: result.stats,
+                    result: QueryResult::GetProviders(Ok(
+                        GetProvidersOk::FinishedWithNoAdditionalRecord {
+                            closest_peers: result.peers.collect(),
+                        },
+                    )),
+                    step,
                 })
             }
 
@@ -1302,65 +1364,41 @@ where
                         ..
                     },
             } => match context {
-                AddProviderContext::Publish => Some(KademliaEvent::OutboundQueryCompleted {
+                AddProviderContext::Publish => Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: get_closest_peers_stats.merge(result.stats),
                     result: QueryResult::StartProviding(Ok(AddProviderOk { key })),
+                    step: ProgressStep::first_and_last(),
                 }),
-                AddProviderContext::Republish => Some(KademliaEvent::OutboundQueryCompleted {
+                AddProviderContext::Republish => Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: get_closest_peers_stats.merge(result.stats),
                     result: QueryResult::RepublishProvider(Ok(AddProviderOk { key })),
+                    step: ProgressStep::first_and_last(),
                 }),
             },
 
             QueryInfo::GetRecord {
                 key,
-                records,
-                quorum,
+                mut step,
+                found_a_record,
                 cache_candidates,
             } => {
-                let results = if records.len() >= quorum.get() {
-                    // [not empty]
-                    if quorum.get() == 1 && !cache_candidates.is_empty() {
-                        // Cache the record at the closest node(s) to the key that
-                        // did not return the record.
-                        let record = records.first().expect("[not empty]").record.clone();
-                        let quorum = NonZeroUsize::new(1).expect("1 > 0");
-                        let context = PutRecordContext::Cache;
-                        let info = QueryInfo::PutRecord {
-                            context,
-                            record,
-                            quorum,
-                            phase: PutRecordPhase::PutRecord {
-                                success: vec![],
-                                get_closest_peers_stats: QueryStats::empty(),
-                            },
-                        };
-                        let inner = QueryInner::new(info);
-                        self.queries
-                            .add_fixed(cache_candidates.values().copied(), inner);
-                    }
-                    Ok(GetRecordOk {
-                        records,
-                        cache_candidates,
-                    })
-                } else if records.is_empty() {
+                step.last = true;
+
+                let results = if found_a_record {
+                    Ok(GetRecordOk::FinishedWithNoAdditionalRecord { cache_candidates })
+                } else {
                     Err(GetRecordError::NotFound {
                         key,
                         closest_peers: result.peers.collect(),
                     })
-                } else {
-                    Err(GetRecordError::QuorumFailed {
-                        key,
-                        records,
-                        quorum,
-                    })
                 };
-                Some(KademliaEvent::OutboundQueryCompleted {
+                Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::GetRecord(results),
+                    step,
                 })
             }
 
@@ -1407,23 +1445,21 @@ where
                 };
                 match context {
                     PutRecordContext::Publish | PutRecordContext::Custom => {
-                        Some(KademliaEvent::OutboundQueryCompleted {
+                        Some(KademliaEvent::OutboundQueryProgressed {
                             id: query_id,
                             stats: get_closest_peers_stats.merge(result.stats),
                             result: QueryResult::PutRecord(mk_result(record.key)),
+                            step: ProgressStep::first_and_last(),
                         })
                     }
-                    PutRecordContext::Republish => Some(KademliaEvent::OutboundQueryCompleted {
+                    PutRecordContext::Republish => Some(KademliaEvent::OutboundQueryProgressed {
                         id: query_id,
                         stats: get_closest_peers_stats.merge(result.stats),
                         result: QueryResult::RepublishRecord(mk_result(record.key)),
+                        step: ProgressStep::first_and_last(),
                     }),
                     PutRecordContext::Replicate => {
                         debug!("Record replicated: {:?}", record.key);
-                        None
-                    }
-                    PutRecordContext::Cache => {
-                        debug!("Record cached: {:?}", record.key);
                         None
                     }
                 }
@@ -1440,54 +1476,66 @@ where
             QueryInfo::Bootstrap {
                 peer,
                 mut remaining,
+                mut step,
             } => {
                 let num_remaining = remaining.as_ref().map(|r| r.len().saturating_sub(1) as u32);
 
-                if let Some(mut remaining) = remaining.take() {
-                    // Continue with the next bootstrap query if `remaining` is not empty.
-                    if let Some(target) = remaining.next() {
-                        let info = QueryInfo::Bootstrap {
-                            peer: target.clone().into_preimage(),
-                            remaining: Some(remaining),
-                        };
-                        let peers = self.kbuckets.closest_keys(&target);
-                        let inner = QueryInner::new(info);
-                        self.queries
-                            .continue_iter_closest(query_id, target.clone(), peers, inner);
-                    }
+                // Continue with the next bootstrap query if `remaining` is not empty.
+                if let Some((target, remaining)) =
+                    remaining.take().and_then(|mut r| Some((r.next()?, r)))
+                {
+                    let info = QueryInfo::Bootstrap {
+                        peer: target.clone().into_preimage(),
+                        remaining: Some(remaining),
+                        step: step.next(),
+                    };
+                    let peers = self.kbuckets.closest_keys(&target);
+                    let inner = QueryInner::new(info);
+                    self.queries
+                        .continue_iter_closest(query_id, target.clone(), peers, inner);
+                } else {
+                    step.last = true;
                 }
 
-                Some(KademliaEvent::OutboundQueryCompleted {
+                Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::Bootstrap(Err(BootstrapError::Timeout {
                         peer,
                         num_remaining,
                     })),
+                    step,
                 })
             }
 
             QueryInfo::AddProvider { context, key, .. } => Some(match context {
-                AddProviderContext::Publish => KademliaEvent::OutboundQueryCompleted {
+                AddProviderContext::Publish => KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::StartProviding(Err(AddProviderError::Timeout { key })),
+                    step: ProgressStep::first_and_last(),
                 },
-                AddProviderContext::Republish => KademliaEvent::OutboundQueryCompleted {
+                AddProviderContext::Republish => KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::RepublishProvider(Err(AddProviderError::Timeout { key })),
+                    step: ProgressStep::first_and_last(),
                 },
             }),
 
-            QueryInfo::GetClosestPeers { key } => Some(KademliaEvent::OutboundQueryCompleted {
-                id: query_id,
-                stats: result.stats,
-                result: QueryResult::GetClosestPeers(Err(GetClosestPeersError::Timeout {
-                    key,
-                    peers: result.peers.collect(),
-                })),
-            }),
+            QueryInfo::GetClosestPeers { key, mut step } => {
+                step.last = true;
+
+                Some(KademliaEvent::OutboundQueryProgressed {
+                    id: query_id,
+                    stats: result.stats,
+                    result: QueryResult::GetClosestPeers(Err(GetClosestPeersError::Timeout {
+                        key,
+                        peers: result.peers.collect(),
+                    })),
+                    step,
+                })
+            }
 
             QueryInfo::PutRecord {
                 record,
@@ -1505,16 +1553,18 @@ where
                 });
                 match context {
                     PutRecordContext::Publish | PutRecordContext::Custom => {
-                        Some(KademliaEvent::OutboundQueryCompleted {
+                        Some(KademliaEvent::OutboundQueryProgressed {
                             id: query_id,
                             stats: result.stats,
                             result: QueryResult::PutRecord(err),
+                            step: ProgressStep::first_and_last(),
                         })
                     }
-                    PutRecordContext::Republish => Some(KademliaEvent::OutboundQueryCompleted {
+                    PutRecordContext::Republish => Some(KademliaEvent::OutboundQueryProgressed {
                         id: query_id,
                         stats: result.stats,
                         result: QueryResult::RepublishRecord(err),
+                        step: ProgressStep::first_and_last(),
                     }),
                     PutRecordContext::Replicate => match phase {
                         PutRecordPhase::GetClosestPeers => {
@@ -1526,45 +1576,31 @@ where
                             None
                         }
                     },
-                    PutRecordContext::Cache => match phase {
-                        PutRecordPhase::GetClosestPeers => {
-                            // Caching a record at the closest peer to a key that did not return
-                            // a record is never preceded by a lookup for the closest peers, i.e.
-                            // it is a direct query to a single peer.
-                            unreachable!()
-                        }
-                        PutRecordPhase::PutRecord { .. } => {
-                            debug!("Caching record failed: {:?}", err);
-                            None
-                        }
-                    },
                 }
             }
 
-            QueryInfo::GetRecord {
-                key,
-                records,
-                quorum,
-                ..
-            } => Some(KademliaEvent::OutboundQueryCompleted {
-                id: query_id,
-                stats: result.stats,
-                result: QueryResult::GetRecord(Err(GetRecordError::Timeout {
-                    key,
-                    records,
-                    quorum,
-                })),
-            }),
+            QueryInfo::GetRecord { key, mut step, .. } => {
+                step.last = true;
 
-            QueryInfo::GetProviders { key, providers } => {
-                Some(KademliaEvent::OutboundQueryCompleted {
+                Some(KademliaEvent::OutboundQueryProgressed {
+                    id: query_id,
+                    stats: result.stats,
+                    result: QueryResult::GetRecord(Err(GetRecordError::Timeout { key })),
+                    step,
+                })
+            }
+
+            QueryInfo::GetProviders { key, mut step, .. } => {
+                step.last = true;
+
+                Some(KademliaEvent::OutboundQueryProgressed {
                     id: query_id,
                     stats: result.stats,
                     result: QueryResult::GetProviders(Err(GetProvidersError::Timeout {
                         key,
-                        providers,
                         closest_peers: result.peers.collect(),
                     })),
+                    step,
                 })
             }
         }
@@ -2063,10 +2099,32 @@ where
                 let peers = closer_peers.iter().chain(provider_peers.iter());
                 self.discovered(&user_data, &source, peers);
                 if let Some(query) = self.queries.get_mut(&user_data) {
-                    if let QueryInfo::GetProviders { providers, .. } = &mut query.inner.info {
-                        for peer in provider_peers {
-                            providers.insert(peer.node_id);
-                        }
+                    let stats = query.stats().clone();
+                    if let QueryInfo::GetProviders {
+                        ref key,
+                        ref mut providers_found,
+                        ref mut step,
+                        ..
+                    } = query.inner.info
+                    {
+                        *providers_found += provider_peers.len();
+                        let providers = provider_peers.iter().map(|p| p.node_id).collect();
+
+                        self.queued_events
+                            .push_back(NetworkBehaviourAction::GenerateEvent(
+                                KademliaEvent::OutboundQueryProgressed {
+                                    id: user_data,
+                                    result: QueryResult::GetProviders(Ok(
+                                        GetProvidersOk::FoundProviders {
+                                            key: key.clone(),
+                                            providers,
+                                        },
+                                    )),
+                                    step: step.clone(),
+                                    stats,
+                                },
+                            ));
+                        *step = step.next();
                     }
                 }
             }
@@ -2138,40 +2196,34 @@ where
                 user_data,
             } => {
                 if let Some(query) = self.queries.get_mut(&user_data) {
+                    let stats = query.stats().clone();
                     if let QueryInfo::GetRecord {
                         key,
-                        records,
-                        quorum,
+                        ref mut step,
+                        ref mut found_a_record,
                         cache_candidates,
                     } = &mut query.inner.info
                     {
                         if let Some(record) = record {
-                            records.push(PeerRecord {
+                            *found_a_record = true;
+                            let record = PeerRecord {
                                 peer: Some(source),
                                 record,
-                            });
+                            };
 
-                            let quorum = quorum.get();
-                            if records.len() >= quorum {
-                                // Desired quorum reached. The query may finish. See
-                                // [`Query::try_finish`] for details.
-                                let peers = records
-                                    .iter()
-                                    .filter_map(|PeerRecord { peer, .. }| peer.as_ref())
-                                    .cloned()
-                                    .collect::<Vec<_>>();
-                                let finished = query.try_finish(peers.iter());
-                                if !finished {
-                                    debug!(
-                                        "GetRecord query ({:?}) reached quorum ({}/{}) with \
-                                         response from peer {} but could not yet finish.",
-                                        user_data,
-                                        peers.len(),
-                                        quorum,
-                                        source,
-                                    );
-                                }
-                            }
+                            self.queued_events
+                                .push_back(NetworkBehaviourAction::GenerateEvent(
+                                    KademliaEvent::OutboundQueryProgressed {
+                                        id: user_data,
+                                        result: QueryResult::GetRecord(Ok(
+                                            GetRecordOk::FoundRecord(record),
+                                        )),
+                                        step: step.clone(),
+                                        stats,
+                                    },
+                                ));
+
+                            *step = step.next();
                         } else {
                             log::trace!("Record with key {:?} not found at {}", key, source);
                             if let KademliaCaching::Enabled { max_peers } = self.caching {
@@ -2323,6 +2375,7 @@ where
                         {
                             query.on_success(&peer_id, vec![])
                         }
+
                         if self.connected_peers.contains(&peer_id) {
                             self.queued_events
                                 .push_back(NetworkBehaviourAction::NotifyHandler {
@@ -2431,14 +2484,16 @@ pub enum KademliaEvent {
     // is made of multiple requests across multiple remote peers.
     InboundRequest { request: InboundRequest },
 
-    /// An outbound query has produced a result.
-    OutboundQueryCompleted {
+    /// An outbound query has made progress.
+    OutboundQueryProgressed {
         /// The ID of the query that finished.
         id: QueryId,
-        /// The result of the query.
+        /// The intermediate result of the query.
         result: QueryResult,
         /// Execution statistics from the query.
         stats: QueryStats,
+        /// Indicates which event this is, if therer are multiple responses for a single query.
+        step: ProgressStep,
     },
 
     /// The routing table has been updated with a new peer and / or
@@ -2490,6 +2545,37 @@ pub enum KademliaEvent {
     /// See [`Kademlia::kbucket`] for insight into the contents of
     /// the k-bucket of `peer`.
     PendingRoutablePeer { peer: PeerId, address: Multiaddr },
+}
+
+/// Information about progress events.
+#[derive(Debug, Clone)]
+pub struct ProgressStep {
+    /// The index into the event
+    pub count: NonZeroUsize,
+    /// Is this the final event?
+    pub last: bool,
+}
+
+impl ProgressStep {
+    fn first() -> Self {
+        Self {
+            count: NonZeroUsize::new(1).expect("1 to be greater than 0."),
+            last: false,
+        }
+    }
+
+    fn first_and_last() -> Self {
+        let mut first = ProgressStep::first();
+        first.last = true;
+        first
+    }
+
+    fn next(&self) -> Self {
+        assert!(!self.last);
+        let count = NonZeroUsize::new(self.count.get() + 1).expect("Adding 1 not to result in 0.");
+
+        Self { count, last: false }
+    }
 }
 
 /// Information about a received and handled inbound request.
@@ -2558,19 +2644,20 @@ pub type GetRecordResult = Result<GetRecordOk, GetRecordError>;
 
 /// The successful result of [`Kademlia::get_record`].
 #[derive(Debug, Clone)]
-pub struct GetRecordOk {
-    /// The records found, including the peer that returned them.
-    pub records: Vec<PeerRecord>,
-    /// If caching is enabled, these are the peers closest
-    /// _to the record key_ (not the local node) that were queried but
-    /// did not return the record, sorted by distance to the record key
-    /// from closest to farthest. How many of these are tracked is configured
-    /// by [`KademliaConfig::set_caching`]. If the lookup used a quorum of
-    /// 1, these peers will be sent the record as a means of caching.
-    /// If the lookup used a quorum > 1, you may wish to use these
-    /// candidates with [`Kademlia::put_record_to`] after selecting
-    /// one of the returned records.
-    pub cache_candidates: BTreeMap<kbucket::Distance, PeerId>,
+pub enum GetRecordOk {
+    FoundRecord(PeerRecord),
+    FinishedWithNoAdditionalRecord {
+        /// If caching is enabled, these are the peers closest
+        /// _to the record key_ (not the local node) that were queried but
+        /// did not return the record, sorted by distance to the record key
+        /// from closest to farthest. How many of these are tracked is configured
+        /// by [`KademliaConfig::set_caching`]. If the lookup used a quorum of
+        /// 1, these peers will be sent the record as a means of caching.
+        /// If the lookup used a quorum > 1, you may wish to use these
+        /// candidates with [`Kademlia::put_record_to`] after selecting
+        /// one of the returned records.
+        cache_candidates: BTreeMap<kbucket::Distance, PeerId>,
+    },
 }
 
 /// The error result of [`Kademlia::get_record`].
@@ -2588,11 +2675,7 @@ pub enum GetRecordError {
         quorum: NonZeroUsize,
     },
     #[error("the request timed out")]
-    Timeout {
-        key: record::Key,
-        records: Vec<PeerRecord>,
-        quorum: NonZeroUsize,
-    },
+    Timeout { key: record::Key },
 }
 
 impl GetRecordError {
@@ -2722,10 +2805,15 @@ pub type GetProvidersResult = Result<GetProvidersOk, GetProvidersError>;
 
 /// The successful result of [`Kademlia::get_providers`].
 #[derive(Debug, Clone)]
-pub struct GetProvidersOk {
-    pub key: record::Key,
-    pub providers: HashSet<PeerId>,
-    pub closest_peers: Vec<PeerId>,
+pub enum GetProvidersOk {
+    FoundProviders {
+        key: record::Key,
+        /// The new set of providers discovered.
+        providers: HashSet<PeerId>,
+    },
+    FinishedWithNoAdditionalRecord {
+        closest_peers: Vec<PeerId>,
+    },
 }
 
 /// The error result of [`Kademlia::get_providers`].
@@ -2734,7 +2822,6 @@ pub enum GetProvidersError {
     #[error("the request timed out")]
     Timeout {
         key: record::Key,
-        providers: HashSet<PeerId>,
         closest_peers: Vec<PeerId>,
     },
 }
@@ -2847,11 +2934,6 @@ pub enum PutRecordContext {
     /// The context is periodic replication (i.e. without extending
     /// the record TTL) of stored records received earlier from another peer.
     Replicate,
-    /// The context is an automatic write-back caching operation of a
-    /// record found via [`Kademlia::get_record`] at the closest node
-    /// to the key queried that did not return a record. This only
-    /// occurs after a lookup quorum of 1 as per standard Kademlia.
-    Cache,
     /// The context is a custom store operation targeting specific
     /// peers initiated by [`Kademlia::put_record_to`].
     Custom,
@@ -2871,17 +2953,25 @@ pub enum QueryInfo {
         /// yet completed and `Some` with an exhausted iterator
         /// if bootstrapping is complete.
         remaining: Option<vec::IntoIter<kbucket::Key<PeerId>>>,
+        step: ProgressStep,
     },
 
-    /// A query initiated by [`Kademlia::get_closest_peers`].
-    GetClosestPeers { key: Vec<u8> },
+    /// A (repeated) query initiated by [`Kademlia::get_closest_peers`].
+    GetClosestPeers {
+        /// The key being queried (the preimage).
+        key: Vec<u8>,
+        /// Current index of events.
+        step: ProgressStep,
+    },
 
-    /// A query initiated by [`Kademlia::get_providers`].
+    /// A (repeated) query initiated by [`Kademlia::get_providers`].
     GetProviders {
         /// The key for which to search for providers.
         key: record::Key,
-        /// The found providers.
-        providers: HashSet<PeerId>,
+        /// The number of providers found so far.
+        providers_found: usize,
+        /// Current index of events.
+        step: ProgressStep,
     },
 
     /// A (repeated) query initiated by [`Kademlia::start_providing`].
@@ -2905,15 +2995,14 @@ pub enum QueryInfo {
         context: PutRecordContext,
     },
 
-    /// A query initiated by [`Kademlia::get_record`].
+    /// A (repeated) query initiated by [`Kademlia::get_record`].
     GetRecord {
         /// The key to look for.
         key: record::Key,
-        /// The records with the id of the peer that returned them. `None` when
-        /// the record was found in the local store.
-        records: Vec<PeerRecord>,
-        /// The number of records to look for.
-        quorum: NonZeroUsize,
+        /// Current index of events.
+        step: ProgressStep,
+        /// Did we find at least one record?
+        found_a_record: bool,
         /// The peers closest to the `key` that were queried but did not return a record,
         /// i.e. the peers that are candidates for caching the record.
         cache_candidates: BTreeMap<kbucket::Distance, PeerId>,